Match exact Original entity-facing arithmetic
Preserve the explicit SWORD conversion performed by the Original Face(RHElement*) overload before subtracting the viewer elevation. Fractional top-plane elevations can otherwise shift a vector across a 16-sector boundary.

Use the full source literals from SBGeoVector2D::GetSector0to15 for its sine, cosine, and tangent classifier constants, retaining the Original left-associative f32 evaluation.

Document the frame-124 pass-door undefined behavior discovered in the schema-10 trace. The Original pass-door movement direction was uninitialized, so that capture selected point_out from stale heap contents. Original commit d1a6e049 now initializes it from the authored gate direction; the existing recording is intentionally valid only through frame 123 and Rust does not emulate its undefined value.

Verified with cargo fmt --all and cargo build --example original_parity_replay.

+++ crates/robin_engine/src/ai/controller.rs
@@ -3027,7 +3027,12 @@ impl AiController {
         let Some(view) = ctx.entity_view(handle) else {
             return;
         };
-        let elevation_delta = view.elevation - ctx.elevation;
+        // The inline Original overload forwards
+        // `(SWORD)pElement->GetElevation()` to Face(RHposition, SWORD).
+        // Preserve that truncation before subtracting the viewer elevation;
+        // fractional projection-plane Z can move a boundary vector into the
+        // adjacent 16-sector direction.
+        let elevation_delta = (view.elevation as i16) as f32 - ctx.elevation;
         let target_pos = view.position;
         self.face_position_impl(target_pos, ctx, elevation_delta, false);
     }
@@ -3037,7 +3042,7 @@ impl AiController {
         let Some(view) = ctx.entity_view(handle) else {
             return;
         };
-        let elevation_delta = view.elevation - ctx.elevation;
+        let elevation_delta = (view.elevation as i16) as f32 - ctx.elevation;
         self.face_position_impl(view.position, ctx, elevation_delta, true);
     }
